backend/requirements.txt
fastapi==0.104.1
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
uvicorn==0.24.0 